Add support for apt proxy (apt-cacher-ng tested).
This is disabled by default, but configured to use the host machines IP
address ( in nat mode ) for the proxy address. You can configure this
as needed. As it is configured for use during the OS installation the
proxy settings are used for APT even after install. This means you
should either disable the proxy config for the box before packaging or
ensure the vagrant host is running a proxy.